Remove WinRT.WeakReference
